Image fix+

+++ api/index.js
@@ -1,4 +1,4 @@
-const express = require("express");
+const express = require('express');
 const cors = require("cors");
 require("dotenv").config();
 const mongoose = require("mongoose");
@@ -6,8 +6,8 @@ const UserModel = require("./models/User");
 const bcrypt = require("bcryptjs");
 const jwt = require("jsonwebtoken");
 const cookieParser = require("cookie-parser");
-const multer = require("multer");
 const path = require("path");
+const helmet = require('helmet');
 
 const Ticket = require("./models/Ticket");
 
@@ -16,6 +16,12 @@ const app = express();
 const bcryptSalt = bcrypt.genSaltSync(10);
 const jwtSecret = "bsbsfbrnsftentwnnwnwn";
 
+app.use(helmet.contentSecurityPolicy({
+   directives: {
+       defaultSrc: ["'self'"],
+       scriptSrc: ["'self'", "'unsafe-inline'", "'unsafe-eval'"],
+   }
+}));
 app.use(express.json());
 app.use(cookieParser());
 app.use(
@@ -25,21 +31,38 @@ app.use(
    })
 );
 
-mongoose.connect(process.env.MONGO_URL);
+mongoose.connect(process.env.MONGO_URL)
+.then(()=>console.log("DB Connected"))
+.catch(err=>console.log(err));
 
-const storage = multer.diskStorage({
-   destination: (req, file, cb) => {
-      cb(null, "uploads/");
-   },
-   filename: (req, file, cb) => {
-      cb(null, file.originalname);
-   },
+const eventSchema = new mongoose.Schema({
+   owner: String,
+   title: String,
+   description: String,
+   organizedBy: String,
+   eventDate: Date,
+   eventTime: String,
+   location: String,
+   Participants: Number,
+   Count: Number,
+   Income: Number,
+   ticketPrice: Number,
+   Quantity: Number,
+   likes: Number,
+   Comment: [String],
+   imageURL: String // New field for storing image URLs
 });
 
-const upload = multer({ storage });
+const Event = mongoose.model("Event", eventSchema);
 
-app.get("/test", (req, res) => {
-   res.json("test ok");
+app.get('/', (req, res) => {
+   Event.find({})
+   .then((data, err)=>{
+       if(err){
+           console.log(err);
+       }
+       res.render('imagepage',{items: data})
+   })
 });
 
 app.post("/register", async (req, res) => {
@@ -91,9 +114,19 @@ app.get("/profile", (req, res) => {
    const { token } = req.cookies;
    if (token) {
       jwt.verify(token, jwtSecret, {}, async (err, userData) => {
-         if (err) throw err;
-         const { name, email, _id } = await UserModel.findById(userData.id);
-         res.json({ name, email, _id });
+         if (err) {
+            // Handle error gracefully
+            console.error("Error verifying token:", err);
+            res.status(500).json({ error: "Internal server error" });
+            return;
+         }
+         try {
+            const { name, email, _id } = await UserModel.findById(userData.id);
+            res.json({ name, email, _id });
+         } catch (error) {
+            console.error("Error fetching user data:", error);
+            res.status(500).json({ error: "Internal server error" });
+         }
       });
    } else {
       res.json(null);
@@ -104,43 +137,44 @@ app.post("/logout", (req, res) => {
    res.cookie("token", "").json(true);
 });
 
-const eventSchema = new mongoose.Schema({
-   owner: String,
-   title: String,
-   description: String,
-   organizedBy: String,
-   eventDate: Date,
-   eventTime: String,
-   location: String,
-   Participants: Number,
-   Count: Number,
-   Income: Number,
-   ticketPrice: Number,
-   Quantity: Number,
-   image: String,
-   likes: Number,
-   Comment: [String],
-});
+app.post("/createEvent", async (req, res) => {
+   const eventData = req.body; // Extract all data from the request body
 
-const Event = mongoose.model("Event", eventSchema);
-
-app.post("/createEvent", upload.single("image"), async (req, res) => {
    try {
-      const eventData = req.body;
-      eventData.image = req.file ? req.file.path : "";
-      const newEvent = new Event(eventData);
+      // Create a new event object using the extracted data
+      const newEvent = new Event({
+         owner: eventData.owner,
+         title: eventData.title,
+         optional: eventData.optional,
+         description: eventData.description,
+         organizedBy: eventData.organizedBy,
+         eventDate: eventData.eventDate,
+         eventTime: eventData.eventTime,
+         location: eventData.location,
+         ticketPrice: eventData.ticketPrice,
+         likes: eventData.likes,
+         imageURL: eventData.imageURL // Include imageURL if it's in the request body
+      });
+
+      // Save the new event
       await newEvent.save();
+
+      // Send the newly created event in the response
       res.status(201).json(newEvent);
    } catch (error) {
+      console.error("Error creating new event:", error.message);
       res.status(500).json({ error: "Failed to save the event to MongoDB" });
    }
 });
 
+
+
 app.get("/createEvent", async (req, res) => {
    try {
-      const events = await Event.find();
+      const events = await Event.find().sort({ eventDate: -1 });
       res.status(200).json(events);
    } catch (error) {
+      console.error("Error fetching events:", error.message);
       res.status(500).json({ error: "Failed to fetch events from MongoDB" });
    }
 });
@@ -149,8 +183,12 @@ app.get("/event/:id", async (req, res) => {
    const { id } = req.params;
    try {
       const event = await Event.findById(id);
+      if (!event) {
+         return res.status(404).json({ message: "Event not found" });
+      }
       res.json(event);
    } catch (error) {
+      console.error("Error fetching event:", error.message);
       res.status(500).json({ error: "Failed to fetch event from MongoDB" });
    }
 });
@@ -171,18 +209,18 @@ app.post("/event/:eventId", (req, res) => {
          res.json(updatedEvent);
       })
       .catch((error) => {
-         console.error("Error liking the event:", error);
+         console.error("Error liking the event:", error.message);
          res.status(500).json({ message: "Server error" });
       });
 });
 
 app.get("/events", (req, res) => {
-   Event.find()
+   Event.find().sort({ eventDate: -1 })
       .then((events) => {
          res.json(events);
       })
       .catch((error) => {
-         console.error("Error fetching events:", error);
+         console.error("Error fetching events:", error.message);
          res.status(500).json({ message: "Server error" });
       });
 });
@@ -191,8 +229,12 @@ app.get("/event/:id/ordersummary", async (req, res) => {
    const { id } = req.params;
    try {
       const event = await Event.findById(id);
+      if (!event) {
+         return res.status(404).json({ message: "Event not found" });
+      }
       res.json(event);
    } catch (error) {
+      console.error("Error fetching event:", error.message);
       res.status(500).json({ error: "Failed to fetch event from MongoDB" });
    }
 });
@@ -201,8 +243,12 @@ app.get("/event/:id/ordersummary/paymentsummary", async (req, res) => {
    const { id } = req.params;
    try {
       const event = await Event.findById(id);
+      if (!event) {
+         return res.status(404).json({ message: "Event not found" });
+      }
       res.json(event);
    } catch (error) {
+      console.error("Error fetching event:", error.message);
       res.status(500).json({ error: "Failed to fetch event from MongoDB" });
    }
 });
@@ -211,20 +257,24 @@ app.post("/tickets", async (req, res) => {
    try {
       const ticketDetails = req.body;
       const newTicket = new Ticket(ticketDetails);
-      await newTicket.save();
-      return res.status(201).json({ ticket: newTicket });
+      const savedTicket = await newTicket.save();
+      return res.status(201).json({ ticket: savedTicket });
    } catch (error) {
-      console.error("Error creating ticket:", error);
+      console.error("Error creating ticket:", error.message);
       return res.status(500).json({ error: "Failed to create ticket" });
    }
 });
 
 app.get("/tickets/:id", async (req, res) => {
    try {
-      const tickets = await Ticket.find();
-      res.json(tickets);
+      const ticketId = req.params.id;
+      const ticket = await Ticket.findById(ticketId);
+      if (!ticket) {
+         return res.status(404).json({ message: "Ticket not found" });
+      }
+      res.json(ticket);
    } catch (error) {
-      console.error("Error fetching tickets:", error);
+      console.error("Error fetching tickets:", error.message);
       res.status(500).json({ error: "Failed to fetch tickets" });
    }
 });
@@ -237,7 +287,7 @@ app.get("/tickets/user/:userId", (req, res) => {
          res.json(tickets);
       })
       .catch((error) => {
-         console.error("Error fetching user tickets:", error);
+         console.error("Error fetching user tickets:", error.message);
          res.status(500).json({ error: "Failed to fetch user tickets" });
       });
 });
@@ -245,10 +295,14 @@ app.get("/tickets/user/:userId", (req, res) => {
 app.delete("/tickets/:id", async (req, res) => {
    try {
       const ticketId = req.params.id;
+      const ticket = await Ticket.findById(ticketId);
+      if (!ticket) {
+         return res.status(404).json({ message: "Ticket not found" });
+      }
       await Ticket.findByIdAndDelete(ticketId);
       res.status(204).send();
    } catch (error) {
-      console.error("Error deleting ticket:", error);
+      console.error("Error deleting ticket:", error.message);
       res.status(500).json({ error: "Failed to delete ticket" });
    }
 });
@@ -257,3 +311,4 @@ const PORT = process.env.PORT || 4000;
 app.listen(PORT, () => {
    console.log(`Server is running on port ${PORT}`);
 });
+
